BC-4053 - move test mocks
